# [Tests] Speed up some fast pipeline tests