--- a/tests/pipelines/controlnet/test_controlnet_sdxl_img2img.py
+++ b/tests/pipelines/controlnet/test_controlnet_sdxl_img2img.py
@@ -169,7 +169,7 @@ class ControlNetPipelineSDXLImg2ImgFastTests(
         }
 
         return inputs
-    
+
     def test_ip_adapter_single(self):
         expected_pipe_slice = np.array([0.6265, 0.5441, 0.5384, 0.5446, 0.5810, 0.5908, 0.5414, 0.5428, 0.5353])
         return super().test_ip_adapter_single(expected_pipe_slice=expected_pipe_slice)