--- a/tests/pipelines/latent_consistency_models/test_latent_consistency_models.py
+++ b/tests/pipelines/latent_consistency_models/test_latent_consistency_models.py
@@ -107,7 +107,7 @@ class LatentConsistencyModelPipelineFastTests(
             "output_type": "np",
         }
         return inputs
-    
+
     def test_ip_adapter_single(self):
         expected_pipe_slice = np.array([0.1403, 0.5072, 0.5316, 0.1202, 0.3865, 0.4211, 0.5363, 0.3557, 0.3645])
         return super().test_ip_adapter_single(expected_pipe_slice=expected_pipe_slice)