… tests/pipelines/pia/test_pia.py
… tests/pipelines/stable_diffusion/test_stable_diffusion_img2img.py
… tests/pipelines/stable_diffusion/test_stable_diffusion_inpaint.py
… tests/pipelines/stable_diffusion_2/test_stable_diffusion_attend_and_excite.py
… tests/pipelines/stable_diffusion_panorama/test_stable_diffusion_panorama.py
… tests/pipelines/stable_diffusion_xl/test_stable_diffusion_xl.py
… tests/pipelines/stable_diffusion_xl/test_stable_diffusion_xl_adapter.py
… tests/pipelines/stable_diffusion_xl/test_stable_diffusion_xl_img2img.py
… tests/pipelines/stable_diffusion_xl/test_stable_diffusion_xl_inpaint.py
… tests/pipelines/test_pipelines_common.py