# What does this PR do?

Currently, the fast pipeline tests take about 20 minutes to fully execute. https://github.com/huggingface/diffusers/actions/runs/8430950874/artifacts/1358187212 gives an idea. 

Below are the tests that take more than 10 seconds to run:

```bash
slowest durations
20.46s call     tests/pipelines/pia/test_pia.py::PIAPipelineFastTests::test_inference_batch_single_identical
20.17s call     tests/pipelines/animatediff/test_animatediff.py::AnimateDiffPipelineFastTests::test_inference_batch_single_identical
19.71s call     tests/pipelines/animatediff/test_animatediff.py::AnimateDiffPipelineFastTests::test_vae_slicing
19.43s call     tests/pipelines/stable_diffusion_2/test_stable_diffusion_attend_and_excite.py::StableDiffusionAttendAndExcitePipelineFastTests::test_karras_schedulers_shape
14.49s call     tests/pipelines/pia/test_pia.py::PIAPipelineFastTests::test_dict_tuple_outputs_equivalent
14.37s call     tests/pipelines/pia/test_pia.py::PIAPipelineFastTests::test_ip_adapter_single
14.28s call     tests/pipelines/pia/test_pia.py::PIAPipelineFastTests::test_ip_adapter_multi
13.95s call     tests/pipelines/animatediff/test_animatediff.py::AnimateDiffPipelineFastTests::test_save_load_local
13.90s call     tests/pipelines/pia/test_pia.py::PIAPipelineFastTests::test_save_load_optional_components
13.55s call     tests/pipelines/animatediff/test_animatediff.py::AnimateDiffPipelineFastTests::test_dict_tuple_outputs_equivalent
13.50s call     tests/pipelines/pia/test_pia.py::PIAPipelineFastTests::test_save_load_local
13.25s call     tests/pipelines/animatediff/test_animatediff.py::AnimateDiffPipelineFastTests::test_save_load_optional_components
10.60s call     tests/pipelines/audioldm2/test_audioldm2.py::AudioLDM2PipelineFastTests::test_audioldm2_num_waveforms_per_prompt
10.55s call     tests/pipelines/stable_diffusion_xl/test_stable_diffusion_xl_adapter.py::StableDiffusionXLAdapterPipelineFastTests::test_StableDiffusionMixin_component
```

IMO we should try to optimize them a little. This PR is a first attempt at that. I tried a lot but couldn't meaningfully optimize the other tests that fall under this category. LMK if you you folks have any suggestions.