@yiyixuxu 2024-03-26
thanks for starting this! @sayakpaul 

for ip_adapter tests, I wonder if instead of getting a baseline output by running the pipeline without the ip-adapter, https://github.com/huggingface/diffusers/blob/544710ef0fe83c3ea87744c21a550d46e088a3c4/tests/pipelines/test_pipelines_common.py#L251 I think maybe we can skip this and just compare it against a static slice 

same can be applied in other places