@sayakpaul 2024-03-27
@yiyixuxu I made some changes to check against static slices for "test_ip_adapter_single". If the plan looks okay to you, I will do the same for "test_ip_adapter_multi". And then in a future PR, I will tackle the following

> same can be applied in other tests e.g. test_dict_tuple_outputs_equivalent, test_inference_batch_single_identical

I think this will be easier for the reviewers too. 

Have left some comments in line to explain some thoughts. 

LMK.