@DN6 2024-03-27
The block output sizes in the models can also be reduced. A lot of them use 32 and 64. These can be reduce to 4 and 8