@sayakpaul 2024-03-27
> The block output sizes in the models can also be reduced. A lot of them use 32 and 64. These can be reduce to 4 and 8

Agree. Could you take care of them in a separate PR? @DN6