@sayakpaul 2024-03-27
From the CI:

> https://github.com/huggingface/diffusers/actions/runs/8447126617/job/23136985269?pr=7477#step:6:12714

It's passing on the CPU GCP VM I am testing this on. I can increase the tolerance to `9e-4`.