@yiyixuxu 2024-03-29
cc @DN6 for a final review 

we can open this task to the community too! 
> If the plan looks okay to you, I will do the same for "test_ip_adapter_multi". And then in a future PR, I will tackle the following

> same can be applied in other tests e.g. test_dict_tuple_outputs_equivalent, test_inference_batch_single_identical