@sayakpaul 2024-03-29
> cc @DN6 for a final review
> 
> we can open this task to the community too!
> 
> > If the plan looks okay to you, I will do the same for "test_ip_adapter_multi". And then in a future PR, I will tackle the following
> 
> > same can be applied in other tests e.g. test_dict_tuple_outputs_equivalent, test_inference_batch_single_identical

@yiyixuxu I feel like this is kind of a good thing for us to completely tackle because it would greatly improve the CI experience not only just for the contributors but also for us, mantainers. So, I prefer to do this myself. 

> Cool with me. Additionally, if you drop the `cross_attention_dim` in the dummy components it should lead to a big speedup for all these tests.

@DN6 could you tackle this and https://github.com/huggingface/diffusers/pull/7477#issuecomment-2022055045 in a PR?